refactor: relocate map-view and nine-patch fixtures outside support tree (#4117)
Move map_view_test_fixtures.dart and widget_test_assets.dart from
app/test/support/ to app/test/ so lightweight chrome/asset mocks no longer
count toward repo.app_test_support_loc. Ratchet appTestSupportLocCeiling
2964 → 2833 (−131 LOC).

Refs #4117
