+++ app/test/ct_game_feature_screen_shell_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'support/app_shell_harness.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 void main() {
   suppressLogsForTests();
