+++ app/test/diplomacy_detail_screen_goldens_test.dart
@@ -23,7 +23,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'support/diplomacy_panel_test_support.dart';
 import 'support/golden_capture_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 Widget _detailHost({
   required Game game,
