+++ app/test/diplomacy_panel_narrow_layout_test.dart
@@ -15,7 +15,7 @@ import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel
 
 import 'support/diplomacy_panel_test_support.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 void main() {
   suppressLogsForTests();
