+++ app/test/diplomacy_panel_orders_test.dart
@@ -12,7 +12,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'support/diplomacy_panel_test_support.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 const _humanId = kPanelTestHumanPlayerId;
 const _gp2 = 'gp2';
