+++ app/test/diplomacy_panel_part1_test.dart
@@ -12,7 +12,7 @@ import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
 
 import 'support/diplomacy_panel_test_support.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 void main() {
   suppressLogsForTests();
