+++ app/test/diplomacy_relative_power_goldens_test.dart
@@ -30,7 +30,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'support/diplomacy_panel_test_support.dart';
 import 'support/golden_capture_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 /// Themed host that mounts [child] under `AppThemes.editorialMonocle` with the
 /// app localization delegates so the relative-power line resolves real copy
