+++ app/test/game_map_area_event_feed_test.dart
@@ -20,7 +20,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'panel_test_fixtures.dart';
 
 class _MapAreaHost extends StatefulWidget {
