+++ app/test/game_map_players_bar_narrow_test.dart
@@ -36,7 +36,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'support/game_screen_test_support.dart';
 import 'panel_test_fixtures.dart';
