+++ app/test/game_map_selection_prompt_dark_tokens_test.dart
@@ -39,7 +39,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'panel_test_fixtures.dart';
 
 void main() {
