+++ app/test/game_screen_320dp_min_viewport_test.dart
@@ -61,7 +61,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'support/min_viewport_harness.dart';
 import 'support/game_screen_test_support.dart';
 import 'panel_test_fixtures.dart';
