+++ app/test/game_screen_narrow_part1_test.dart
@@ -17,7 +17,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
 import 'support/game_screen_test_support.dart';
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'panel_test_fixtures.dart';
 
 void main() {
