+++ app/test/game_screen_s13_mockup_fidelity_test.dart
@@ -32,7 +32,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'panel_test_fixtures.dart';
 import 'support/game_screen_test_support.dart';
