+++ app/test/map_view_test_fixtures.dart
