+++ app/test/map_view_test_fixtures_test.dart
@@ -7,7 +7,7 @@ import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter_test/flutter_test.dart';
 
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 
 void main() {
   suppressLogsForTests();
