+++ app/test/naval_units_panel_part1_test.dart
@@ -15,7 +15,7 @@ import 'package:colonizethis_app/widgets/ct_transfer_list.dart';
 
 import 'support/naval_units_panel_test_support.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 void main() {
   suppressLogsForTests();
