+++ app/test/naval_units_panel_part3_test.dart
@@ -13,7 +13,7 @@ import 'package:colonizethis_app/widgets/ct_nine_patch_button.dart';
 import 'package:colonizethis_app/widgets/ct_transfer_list.dart';
 
 import 'support/naval_units_panel_test_support.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 const String _mergePort = 'oldWorld|mergeport';
