+++ app/test/player_turn_event_feed_narrow_inset_test.dart
@@ -49,7 +49,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hive/hive.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/map_view_test_fixtures.dart';
+import 'map_view_test_fixtures.dart';
 import 'panel_test_fixtures.dart';
 
 /// Narrow viewport surface size — under `kNarrowBreakpoint` (600 dp) so
