+++ app/test/units_panel_test_support_test.dart
@@ -13,7 +13,7 @@ import 'package:colonizethis_app/features/game/widgets/units/naval/naval_units_p
 import 'support/military_units_panel_test_support.dart';
 import 'support/naval_units_panel_test_support.dart';
 import 'panel_test_fixtures.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 
 void main() {
   suppressLogsForTests();
