+++ app/test/widgetbook_diplomacy_detail_screen_stories_test.dart
@@ -19,7 +19,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:widgetbook/widgetbook.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 const String _kFolder = 'Diplomacy Detail Screen';
