+++ app/test/widgetbook_diplomacy_panel_empty_state_test.dart
@@ -26,7 +26,7 @@ import 'package:colonizethis_app_l10n/l10n/l10n.dart';
 import 'package:widgetbook_host/catalogs/catalog.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 void main() {
