+++ app/test/widgetbook_diplomacy_panel_mobile_viewport_test.dart
@@ -27,7 +27,7 @@ import 'package:widgetbook/widgetbook.dart';
 import 'package:colonizethis_app/features/game/widgets/diplomacy/diplomacy_panel.dart';
 import 'package:widgetbook_host/catalogs/catalog.dart';
 
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 void main() {
