+++ app/test/widgetbook_diplomacy_standing_chips_stories_test.dart
@@ -23,7 +23,7 @@ import 'package:colonizethis_app_l10n/l10n/l10n.dart';
 import 'package:widgetbook_host/catalogs/catalog.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 const String _kFolder = 'Diplomatic Standing Chips';
