+++ app/test/widgetbook_main_menu_stories_editorial_monocle_test.dart
@@ -19,7 +19,7 @@ import 'package:colonizethis_app/widgets/main_menu.dart';
 import 'package:widgetbook_host/catalogs/catalog.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 /// Normative inventory for issue #2860 S6 — renaming or removing a story
