+++ app/test/widgetbook_province_overlay_mobile_viewport_test.dart
@@ -26,7 +26,7 @@ import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_host/catalogs/catalog.dart';
 
 import 'support/app_shell_harness.dart';
-import 'support/widget_test_assets.dart';
+import 'widget_test_assets.dart';
 import 'support/widgetbook_test_harness.dart';
 
 void main() {
