+++ tool/check_app_test_support_loc.dart
@@ -13,8 +13,9 @@ import 'package:path/path.dart' as p;
 const String appTestSupportRelativeDir = 'app/test/support';
 
 /// Post-#4035 densify support-tree ceiling (physical LOC).
-/// Ratcheted after wave-9 slice F debug-handler fixtures relocate outside support tree (Refs #4117).
-const int appTestSupportLocCeiling = 2964;
+/// Ratcheted after wave-9 slice G map-view + nine-patch asset fixtures relocate
+/// outside support tree (Refs #4117).
+const int appTestSupportLocCeiling = 2833;
 
 /// Counts physical lines of all `*.dart` files under [dir].
 int countAppTestSupportPhysicalLoc(Directory dir) {
